+++ src/FluxFlow.Components.Assertions.Composition/AssertionsCompositionPortNames.cs
@@ -5,8 +5,4 @@ public static class AssertionsCompositionPortNames
     public const string Input = "Input";
 
     public const string Output = "Output";
-
-    public const string Passed = "Passed";
-
-    public const string Failed = "Failed";
 }